# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.AdminVO:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      phone:
        type: string
      role_names:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
      status:
        type: integer
      status_name:
        type: string
      username:
        type: string
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 32
        minLength: 6
        type: string
      old_password:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
  service.CreateAdminRequest:
    properties:
      email:
        maxLength: 128
        type: string
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      phone:
        maxLength: 20
        type: string
      role_ids:
        items:
          type: integer
        type: array
      status:
        type: integer
      username:
        maxLength: 20
        minLength: 4
        type: string
    required:
    - password
    - username
    type: object
...
  service.LoginRequest:
    properties:
      password:
        maxLength: 32
        minLength: 6
        type: string
      username:
        maxLength: 32
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  service.LoginResponse:
    properties:
      admin:
        $ref: '#/definitions/model.AdminVO'
      expires_in:
        type: integer
      token:
        type: string
    type: object
...
  service.ResetAdminPasswordRequest:
    properties:
      password:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - password
    type: object
...
  service.UpdateAdminRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@huaan-medical.com
    name: API Support
  description: 华安医疗门诊预约系统后端API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 华安医疗预约系统 API
  version: "1.0"
paths:
  /api/admin/admins:
    get:
      consumes:
      - application/json
      description: 分页查询管理员列表
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      - description: 关键词(用户名/昵称/手机号/邮箱)
        in: query
        name: keyword
        type: string
      - description: 状态 0禁用 1启用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.AdminVO'
                        type: array
                    type: object
              type: object
      security:
      - BearerAdmin: []
      summary: 管理员列表
      tags:
      - 管理员管理
    post:
      consumes:
      - application/json
      description: 创建管理员并可选绑定角色
      parameters:
      - description: 管理员信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AdminVO'
              type: object
      security:
      - BearerAdmin: []
      summary: 创建管理员
      tags:
      - 管理员管理
  /api/admin/admins/{id}:
    put:
      consumes:
      - application/json
      description: 更新管理员信息与角色
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      - description: 管理员信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AdminVO'
              type: object
      security:
      - BearerAdmin: []
      summary: 更新管理员
      tags:
      - 管理员管理
  /api/admin/admins/{id}/password:
    put:
      consumes:
      - application/json
      description: 重置管理员登录密码
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      - description: 密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ResetAdminPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAdmin: []
      summary: 重置管理员密码
      tags:
      - 管理员管理
...
  /api/admin/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录管理员的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AdminVO'
              type: object
      security:
      - Bearer: []
      summary: 获取当前管理员信息
      tags:
      - 管理员
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: 管理员账号密码登录
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.LoginResponse'
              type: object
      summary: 管理员登录
      tags:
      - 管理员
...
  /api/admin/password:
    put:
      consumes:
      - application/json
      description: 修改当前管理员密码
      parameters:
      - description: 密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: 修改密码
      tags:
      - 管理员
...
securityDefinitions:
  Bearer:
    description: 请在Token前加上"Bearer "前缀
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"